# [WIP] Added a simple performance test with the help of Benchmark.js