--- a/package.json
+++ b/package.json
@@ -20,6 +20,7 @@
   "author": "Emil Englesson",
   "license": "MIT",
   "devDependencies": {
+    "benchmark": "^2.1.4",
     "mocha": "^5.2.0"
   }
 }