--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@limeeng/semaphore",
-  "version": "0.5.7",
+  "version": "0.5.8",
   "description": "Semaphores, with promises. Limit access to resources in a safe and easy manner",
   "main": "index.js",
   "repository": {