This is a simple benchmarking test that will help measure performance when implementing #3. 

Performance output right now:
```
λ node test\performance\perf.js
Executing tests on:
==============
Name:         Node.js
Version:      10.5.0
Product:      null
Manufacturer: null
Layout:       null
OS:           Win32 64-bit
Description:  Node.js 10.5.0 on Win32 64-bit
==============
1/6 completed: Semaphore(10) -> 10 promises x 132,377 ops/sec ±0.52% (87 runs sampled)
2/6 completed: Semaphore(10) -> 100 promises x 10,571 ops/sec ±0.40% (89 runs sampled)
3/6 completed: Semaphore(10) -> 1000 promises x 373 ops/sec ±0.26% (88 runs sampled)
4/6 completed: Semaphore(10) -> 10000 promises x 6.47 ops/sec ±0.23% (35 runs sampled)
5/6 completed: Semaphore(10) -> 100000 promises x 0.03 ops/sec ±1.98% (5 runs sampled)
^C
```

Signed-off-by: Emil Englesson <englesson.emil@gmail.com>